Company Employees Project
In the company employees project, the salaries, departments and information of all employees were kept in the program. MYSQL database was connected to the program, all information in the database was updated integrated with the program, ids were automatically sorted with MYSQL database, administrator login connection was made with MYSQL database, administrator username and password was added. Program entry is under control. Dynamically running the search in my program, adding and removing employees; Employee information updates were made.